src/main/java/com/glowbyte/task/model/CarComponentAction.java
package com.glowbyte.task.model;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.glowbyte.task.exception.InvariantException;
import lombok.AccessLevel;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;

@NoArgsConstructor(access = AccessLevel.PROTECTED)
@Data
public class CarComponentAction implements Serializable {

    private String name;

    @JsonCreator
    public CarComponentAction(String name) {
        setName(name);
    }

    public void setName(String name) {
        if (name == null || name.isBlank()) {
            throw new InvariantException(400, "Car component action's name must be presented");
        }
        this.name = name;
    }
}
